% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Download a single FAO bulk download file}
\usage{
get_data(DatasetCode = "QP")
}
\value{
a data_frame.
}
\description{
Downloads an individual FAO bulk download file based on \code{DatasetCode}
}
\examples{
 \dontrun{
 get_data(DatasetCode = "QP")
 }

}
\author{
Markus Kainu <muuankarski@kapsi.fi>
}
