% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datalist.R
\name{get_datalist}
\alias{get_datalist}
\title{Download Database description}
\usage{
get_datalist()
}
\value{
a data_frame.
}
\description{
Downloads a description file containing information on available bulk download files
}
\details{
The description file XML URL is \url{http://fenixservices.fao.org/faostat/static/bulkdownloads/datasets_E.xml}.
}
\examples{
 \dontrun{
 get_datalist()
 }

}
\author{
Markus Kainu <muuankarski@kapsi.fi>
}
